==> all-features/1-partition.ipynb <==
# Assign positives and negatives 

import itertools
import random
import sys
import json

import py2neo
import pandas

## Startup neo4j and connections

with open('servers.json') as read_file:
    instances = json.load(read_file)

for instance in instances:
    neo = py2neo.Graph(
        host="localhost", bolt_port=instance['port'],
        http_port=instance['port']+1, bolt=True
    )
    
    instance['py2neo'] = neo
    
    if instance['name'] == 'rephetio-v2.0':
        neo_unperm = instance['py2neo']

hetnets = [x['name'] for x in instances]
hetnets

## Create partitions

def to_df(record_list):
    """Convert a py2neo RecordList to a dataframe"""
    return pandas.DataFrame(record_list.data())

compound_df = pandas.read_table('../summary/compounds.tsv')
disease_df = pandas.read_table('../summary/diseases.tsv')

## All potential pairs

'{} compounds × {} diseases = {} pairs'.format(
    len(disease_df), len(compound_df), len(disease_df) * len(compound_df))

## Nonzero prior pairs

compound_df.query("treats > 0")["chemical_id"].nunique()

disease_df.query("treats > 0")["disease_id"].nunique()

nonzero_prior_pairs = set(itertools.product(
    compound_df.query("treats > 0").chemical_id,
    disease_df.query("treats > 0").disease_id)
)

len(nonzero_prior_pairs)

---

## Read training data for cross validation

with open("../../crossval_idx.txt", "r") as fin:
    train_idx = int(fin.read().strip())

train_idx

train = pandas.read_csv("../../../crossval/data/training/training_piece{}.tsv".format(train_idx), sep = '\t')

train.shape

def df_to_pairs(df):
    return set(zip(df["chemical_id"], df["disease_id"]))    

train_pairs = df_to_pairs(train)

nonzero_prior_pairs <= train_pairs

There are some potential pairs which are present in the holdout set, so we will remove these from the potential negative training examples.

nonzero_prior_pairs &= train_pairs

len(nonzero_prior_pairs)

nonzero_prior_pairs <= train_pairs

The data has now been sanitized so that no relations which appear in the holdout set are ever seen prior to making predictions.

---

## Split the data

indication_query = '''
MATCH (compound:Compound)-[treatment:TREATS_CtD]->(disease:Disease)
RETURN
  compound.identifier AS chemical_id,
  disease.identifier AS disease_id
ORDER BY
  chemical_id, disease_id
'''

def partition(neo):
    """
    Extract negative and positive compound-disease pairs from a py2neo.Graph.
    """
    
    # query hetnet for all positive drug-treat-disease examples
    # query hetnet for all potential drug-disease pairs
    # randomly sample 4x the number of non-positive examples for use as negative examples

    
    # Use TREATS_CtD as positives
    # different from training data because edges were permuted
    indication_df = to_df(neo.run(indication_query))
    
    positives = set(zip(indication_df.chemical_id, indication_df.disease_id))

    # Use nonzero-prior pairs excluding non-negatives as negatives
    negatives = nonzero_prior_pairs - positives
    
    assert negatives <= train_pairs, "Bad training data"
    
    
    negatives = random.sample(negatives, k=len(positives) * 4)
    rows = list()
    for status, pairs in (0, negatives), (1, positives):
        for drug, disease in pairs:
            rows.append((drug, disease, status))
    df = pandas.DataFrame(rows, columns=['chemical_id', 'disease_id', 'status'])
    df = df.sort_values(['disease_id', 'status', 'chemical_id'])
    return df

part_dfs = list()
for seed, instance in enumerate(instances):
    random.seed(seed, version=2)
    part_df = partition(instance['py2neo'])
    part_df.insert(0, 'hetnet', instance['name'])
    part_dfs.append(part_df)
part_df = pandas.concat(part_dfs)

part_df.head()

part_df.groupby("hetnet")["status"].value_counts()

## Create an even-matrix for unpermuted observations
template_df = pandas.DataFrame(list(itertools.product(hetnets, compound_df.chemical_id)), columns=['hetnet', 'chemical_id'])
unperm_pair_df = part_df.query("hetnet == 'rephetio-v2.0'")[['chemical_id', 'disease_id']]
unperm_pair_df = template_df.merge(unperm_pair_df)
unperm_pair_df = unperm_pair_df.merge(part_df, how='left')
unperm_pair_df = unperm_pair_df[unperm_pair_df.status.isnull()]
unperm_pair_df.status = unperm_pair_df.status.fillna(0).astype(int)
unperm_pair_df.head(2)

part_df['primary'] = 1
unperm_pair_df['primary'] = 0
full_part_df = pandas.concat([part_df, unperm_pair_df])
full_part_df = full_part_df.sort_values(['chemical_id', 'disease_id', 'hetnet'])
full_part_df.head(5)

full_part_df.to_csv('data/partitions.tsv', sep='\t', index=False)

# Number of hetnet-compound-disease pairs
len(full_part_df)

# Number of positives and negatives per purpose
pandas.crosstab(full_part_df.primary, full_part_df.status)

# Number of positives and negatives per hetnet
pandas.crosstab(full_part_df.hetnet, full_part_df.status)
